+++ tests/unit/unit_t_db_base.py
@@ -134,7 +134,7 @@ class UnitTestDbBase(unittest.TestCase):
         """
         self.set_up_client()
 
-    def set_up_client(self, auto_connect=False):
+    def set_up_client(self, auto_connect=False, encoder=None):
         if os.environ.get('RUN_CLOUDANT_TESTS') is None:
             admin_party = False
             if (os.environ.get('ADMIN_PARTY') and
@@ -148,7 +148,8 @@ class UnitTestDbBase(unittest.TestCase):
                 self.pwd,
                 admin_party,
                 url=self.url,
-                connect=auto_connect
+                connect=auto_connect,
+                encoder=encoder
             )
         else:
             self.account = os.environ.get('CLOUDANT_ACCOUNT')
@@ -162,7 +163,8 @@ class UnitTestDbBase(unittest.TestCase):
                 self.pwd,
                 url=self.url,
                 x_cloudant_user=self.account,
-                connect=auto_connect
+                connect=auto_connect,
+                encoder=encoder
             )